requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yt-dlp==2025.1.15
openai-whisper==20231117
transformers==4.44.2
torch==2.2.2
pydantic==2.9.2
numpy<2
